+++ NEWS.md
@@ -1,5 +1,8 @@
 # CHANGES IN DT VERSION 0.25
 
+## NEW FEATURES
+
+- Server-side filtering is now able to be ran as a `future` via the **promises** package (@andyquinterom, #1011). See the argument `future` on the help page `?DT::renderDataTable` for more info.
 
 # CHANGES IN DT VERSION 0.24